Fix bad check if content is placeholder
The `lastContent == foundContent` check, last touched in 9c6422fe, is
terminally broken, as it would always be false. `foundContent` is
mutated when a new card load task is started in `onSearchFinished()`,
which is *before* the aforementioned check.

The code prior to 9c6422fe was checking against the two static reused
placeholder drawables which was the correct check to apply, and this
commit reverts to using a variant of that check.