Add debug language to help with localisation efforts
The idea is to allow a developer to immediately see which text on a
component or screen has already got localisation support. It can be a
bit of a challenge to see this when creating a new component that
doesn't yet have any translations populated.

Curious to hear thoughts on this. I could see it working very well as a
visual tests checkbox (implemented at o!f side), potentially in addition
to having this at the game level, or replacing this PR.