Fix joystick settings changing enabled state of config level bindables
Just copied what we had in place for the mouse sensitivity setting. For
tablet settings I believe I avoided this by just hiding the settings
when not enabled. Might be a better way forward, but this is simplest
for now.

Run `TestSceneSettingsPanel` twice to get a crash without this change.