Fix unobserved exceptions due to unimplemented beatmap converter
`TestSceneStatisticsPanel` intends to check the operation of statistics
panels using dummy ruleset classes. However, `StatisticsPanel` relies on
being able to retrieve the playable beatmap, which requires a converter.
One was not provided by the dummy rulesets, therefore the retrieval
would fail with an unobserved exception.

To fix, add a barebones converter implementation that is enough for the
test to pass.